app/scan/network_scanner.py
import nmap
import socket
import logging
from typing import List, Dict, Any
import asyncio
from concurrent.futures import ThreadPoolExecutor

logger = logging.getLogger(__name__)


class NetworkScanner:
    """Network vulnerability scanner using Nmap"""
    
    def __init__(self):
        self.nm = nmap.PortScanner()
        self.executor = ThreadPoolExecutor(max_workers=3)
    
    async def discover_assets(self, target: str = "127.0.0.1") -> List[Dict[str, Any]]:
        """
        Discover network assets
        
        Args:
            target: IP address or range to scan (default: localhost for Windows compatibility)
        
        Returns:
            List of discovered assets
        """
        assets = []
        
        try:
            # Run nmap scan in thread pool to avoid blocking
            loop = asyncio.get_event_loop()
            await loop.run_in_executor(
                self.executor,
                self.nm.scan,
                target,
                '22-443',
                '-sV -T4'
            )
            
            for host in self.nm.all_hosts():
                if self.nm[host].state() == 'up':
                    asset = {
                        'ip': host,
                        'hostname': self.nm[host].hostname() or 'Unknown',
                        'state': self.nm[host].state(),
                        'ports': [],
                        'services': []
                    }
                    
                    # Get port and service information
                    for proto in self.nm[host].all_protocols():
                        ports = self.nm[host][proto].keys()
                        for port in ports:
                            port_info = self.nm[host][proto][port]
                            asset['ports'].append({
                                'port': port,
                                'state': port_info['state'],
                                'service': port_info.get('name', 'unknown'),
                                'version': port_info.get('version', ''),
                                'product': port_info.get('product', '')
                            })
                            
                            if port_info.get('product'):
                                asset['services'].append({
                                    'name': port_info.get('name', 'unknown'),
                                    'product': port_info.get('product', ''),
                                    'version': port_info.get('version', ''),
                                    'extrainfo': port_info.get('extrainfo', '')
                                })
                    
                    assets.append(asset)
            
            logger.info(f"Discovered {len(assets)} network assets")
            
        except Exception as e:
            logger.error(f"Asset discovery failed: {e}")
            # Return minimal localhost info for demo purposes
            assets = [{
                'ip': '127.0.0.1',
                'hostname': socket.gethostname(),
                'state': 'up',
                'ports': [
                    {'port': 80, 'state': 'open', 'service': 'http', 'version': '', 'product': ''},
                    {'port': 443, 'state': 'open', 'service': 'https', 'version': '', 'product': ''}
                ],
                'services': []
            }]
        
        return assets
    
    async def quick_scan(self, target: str = "127.0.0.1") -> Dict[str, Any]:
        """
        Perform quick network scan for common ports
        
        Args:
            target: Target IP or hostname
        
        Returns:
            Scan results dictionary
        """
        result = {
            'target': target,
            'scan_type': 'quick',
            'open_ports': [],
            'status': 'completed'
        }
        
        try:
            loop = asyncio.get_event_loop()
            await loop.run_in_executor(
                self.executor,
                self.nm.scan,
                target,
                '21,22,23,25,53,80,110,143,443,445,3306,3389,5432,8080,8443'
            )
            
            if target in self.nm.all_hosts():
                for proto in self.nm[target].all_protocols():
                    ports = self.nm[target][proto].keys()
                    for port in ports:
                        port_info = self.nm[target][proto][port]
                        if port_info['state'] == 'open':
                            result['open_ports'].append({
                                'port': port,
                                'service': port_info.get('name', 'unknown'),
                                'product': port_info.get('product', ''),
                                'version': port_info.get('version', '')
                            })
        
        except Exception as e:
            logger.error(f"Quick scan failed: {e}")
            result['status'] = 'error'
            result['error'] = str(e)
        
        return result
